requirements.txt
# this file holds all requirements to run the package, excluding optional requirements
# make sure all packages are pinned
googleapis-common-protos==1.63.0
google-cloud-bigquery==3.18.0
google-cloud-storage==2.16.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
arrow==1.3.0
db-dtypes==1.2.0
pandas==2.0.3
polars==0.20.20
pandas-gbq==0.22.0
sqlparse==0.5.0
python-snappy==0.7.1
Pygments==2.17.2
Click==8.1.7
six==1.16.0
pyarrow==15.0.2
fuzzywuzzy==0.18.0
pyyaml==6.0.1
fastavro==1.9.4
ipython==8.12.3
jinja2==3.1.3
python-Levenshtein==0.25.1
tqdm==4.66.2
plaid_python==14.0.0
notebook==7.0.7